Further reorganisation of the  app structure. Created 'routes' folder.
