+++ src/components/routes/app/results/ItemTiles.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react'
-import ItemTile from 'ItemTile'
+import ItemTile from './ItemTile'
 
 const ItemTiles = ({items}) => {
   let tiles = items.map(
