+++ src/components/routes/app/search/FilterByDifficulty.js
