+++ src/components/routes/app/search/index.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react'
-import { selectSearchFilter, setSearchResults } from '../../actions'
+import { selectSearchFilter, setSearchResults } from '../../../../actions'
 import { connect } from 'react-redux'
-import {RESOURCE_TYPE, DIFFICULTY} from '../../constants/FilterCategories'
+import {RESOURCE_TYPE, DIFFICULTY} from '../../../../constants/FilterCategories'
 import FilterByResource from './FilterByResource'
 import FilterByDifficulty from './FilterByDifficulty'
 import FilterByKeyword from './FilterByKeyword'
