+++ src/index.js
@@ -4,9 +4,9 @@ import ReactDOM from 'react-dom'
 import { Provider } from 'react-redux'
 import { createStore, applyMiddleware } from 'redux'
 import { Router, Route, browserHistory } from 'react-router'
-import App from 'App'
-import Admin from 'Admin'
-import ItemEdit from 'ItemEdit'
+import App from './components/routes/app/App'
+import Admin from './components/routes/admin/Admin'
+import ItemEdit from './components/routes/itemEdit/ItemEdit'
 import rootReducer from 'index'
 import '../public/assets/css/bootstrap.css'
 import '../public/assets/css/grid.css'
